Move NoServiceConfiguredException from styx-api to styx-server module.
