+++ components/proxy/src/main/java/com/hotels/styx/proxy/RouteHandlerAdapter.java
@@ -19,7 +19,7 @@ import com.hotels.styx.api.HttpHandler;
 import com.hotels.styx.api.HttpInterceptor;
 import com.hotels.styx.api.HttpRequest;
 import com.hotels.styx.api.HttpResponse;
-import com.hotels.styx.api.NoServiceConfiguredException;
+import com.hotels.styx.server.NoServiceConfiguredException;
 import com.hotels.styx.api.StyxObservable;
 import com.hotels.styx.server.HttpRouter;
