+++ components/server/src/main/java/com/hotels/styx/server/NoServiceConfiguredException.java
@@ -13,7 +13,7 @@
   See the License for the specific language governing permissions and
   limitations under the License.
  */
-package com.hotels.styx.api;
+package com.hotels.styx.server;
 
 import static java.lang.String.format;
